# java/socialapp/src/main/java/com/contoso/socialapp/dto/CreatePostRequest.java
package com.contoso.socialapp.dto;

import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CreatePostRequest {
    private String username;
    private String content;
}